+++ src/components/elecar/list/detail/DetailTable.vue
