+++ src/components/welding/list/detail/DetailTable.vue
